+++ src/com/lucienrowan/financemanager/db/SqliteManager.java
@@ -1,6 +1,7 @@
 package com.lucienrowan.financemanager.db;
 
 import javax.swing.*;
+import java.io.File;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
@@ -8,6 +9,17 @@ import java.sql.SQLException;
 public class SqliteManager {
     public static Connection getConnection(String filePath) {
         try {
+            File file = new File(filePath);
+
+            if (file.exists() && file.isDirectory()) {
+                filePath += "/fiance_db.sqlite";
+                file = new File(filePath);
+            }
+
+            if (file.createNewFile()) {
+                // TODO: Report that file was created successfully
+            }
+
             return DriverManager.getConnection("jdbc:sqlite:" + filePath);
         } catch (SQLException e) {
             JOptionPane.showMessageDialog(null, e.getLocalizedMessage(), "Error",